wip: cleanup rds
